refactor: replace LocalStack references with Substrate throughout codebase
- pkg/aws/s3: extract LocalStack-specific test to new
  adaptive_transporter_integration_test.go (//go:build integration)
  using Substrate in-process emulator via existing substrateURL from
  integration_test.go; remove Docker-dependent helpers from base file
- pkg/aws/config: clarify IsLocalStackConfig() deprecation comment
- pkg/aws/config (test): rename LocalStack-named variables to generic names
- pkg/multiregion: update emulator compatibility comments
- cmd/cargoship/cmd: update setupS3Client comment
